remove user app for optimization
