+++ careernest/urls.py
@@ -3,5 +3,4 @@ from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('user/', include('user.urls'))
 ]
